cleanup rm -rf protection and userdata as root

+++ install.sh
@@ -8,22 +8,32 @@ adb shell mkdir /tmp/system_a
 adb shell mkdir /tmp/system_b
 adb shell mkdir /tmp/userdata
 
-adb shell mount /dev/block/bootdevice/by-name/system_a /tmp/system_a
-adb shell mount /dev/block/bootdevice/by-name/system_b /tmp/system_b
-adb shell mount /dev/block/bootdevice/by-name/userdata /tmp/userdata
-adb shell fallocate -l 42G /tmp/userdata/system.img
-adb shell mkfs.ext4 /tmp/userdata/system.img
-adb shell mkdir /tmp/userdata/system
-adb shell mount /tmp/userdata/system.img /tmp/userdata/system
+if false
+then
+	adb shell mount /dev/block/bootdevice/by-name/system_a /tmp/system_a
+	adb shell mount /dev/block/bootdevice/by-name/system_b /tmp/system_b
+	adb shell mount /dev/block/bootdevice/by-name/userdata /tmp/userdata
+	adb shell fallocate -l 42G /tmp/userdata/system.img
+	adb shell mkfs.ext4 /tmp/userdata/system.img
+	adb shell mkdir /tmp/userdata/system
+	adb shell mount /tmp/userdata/system.img /tmp/userdata/system
 
-cat rootfs_out/first_stage.tar | adb shell 'busybox tar -C /tmp/system_a -xv'
-cat rootfs_out/first_stage.tar | adb shell 'busybox tar -C /tmp/system_b -xv'
-cat rootfs_out/rootfs.tar | adb shell 'busybox tar -C /tmp/userdata/system -xv'
+	cat rootfs_out/first_stage.tar | adb shell 'busybox tar -C /tmp/system_a -xv'
+	cat rootfs_out/first_stage.tar | adb shell 'busybox tar -C /tmp/system_b -xv'
+	cat rootfs_out/rootfs.tar | adb shell 'busybox tar -C /tmp/userdata/system -xv'
 
-adb shell umount /tmp/system_a
-adb shell umount /tmp/system_b
-adb shell umount /tmp/userdata/system
-adb shell umount /tmp/userdata
+	adb shell umount /tmp/system_a
+	adb shell umount /tmp/system_b
+	adb shell umount /tmp/userdata/system
+	adb shell umount /tmp/userdata
+else
+	adb shell mount /dev/block/bootdevice/by-name/userdata /tmp/userdata
+
+	cat rootfs_out/rootfs.tar | adb shell 'busybox tar -C /tmp/userdata -xv'
+
+	adb shell umount /tmp/userdata	
+fi
+
 adb shell sync
 
 cat boot_out/boot_recovery.img | adb shell 'dd of=/dev/block/bootdevice/by-name/boot_a'
